Speed up by factor of 3.6 (27.8% runtime)
The profiler VisualVM showed that 90% of the runtime is consumed in
method PlayingField.canBePlayed(..). I found out that mapping the stack
of moves to pieces played and filtering the pieces via Java 8 streams,
maps and filters was really slow. Just keeping an additional set of
pieces played speeds up the whole thing considerably.

Timing stats:
  - 3 rows, 117 solutions found, 117.5 s instead of 423.3 s